+++ android/src/main/java/com/odehbros/flutter_file_downloader/MethodCallHandlerImpl.java
@@ -52,6 +52,16 @@ public class MethodCallHandlerImpl implements MethodChannel.MethodCallHandler {
         return null;
     }
 
+    public void removeStoreHelper(final long id) {
+        final String toFind = String.valueOf(id);
+        for (final String key : stored.keySet()) {
+            if ((toFind + "").equals(stored.get(key).id + "")) {
+                stored.remove(key);
+                return;
+            }
+        }
+    }
+
     @Override
     public void onMethodCall(@NonNull MethodCall call, @NonNull MethodChannel.Result result) {
         final StoreHelper helper = new StoreHelper(call, result);
@@ -159,13 +169,12 @@ public class MethodCallHandlerImpl implements MethodChannel.MethodCallHandler {
                 .setShowNotifications(notifications)
                 .setRequestHeaders(requestHeaders)
                 .setDownloadDestination(lastDestination)
+                .setStoreHelper(helper)
                 .setCallbacks(new DownloadCallbacks() {
                     @Override
                     public void onIDReceived(long id) {
                         super.onIDReceived(id);
                         tasks.put(id, this);
-
-                        final String onProgress = helper.call.argument("onidreceived");
 
                         Map<String, Object> args = new HashMap();
 
@@ -180,9 +189,6 @@ public class MethodCallHandlerImpl implements MethodChannel.MethodCallHandler {
                     public void onProgress(double progress) {
                         super.onProgress(progress);
 
-                        final String onProgress = helper.call.argument("onprogress");
-
-                        if (TextUtils.isEmpty(onProgress)) return;
                         Map<String, Object> args = new HashMap();
 
                         args.put("id", id);
@@ -195,9 +201,6 @@ public class MethodCallHandlerImpl implements MethodChannel.MethodCallHandler {
                     public void onProgress(String name, double progress) {
                         super.onProgress(name, progress);
 
-                        final String onProgressWithName = helper.call.argument("onprogress_named");
-
-                        if (TextUtils.isEmpty(onProgressWithName)) return;
                         Map<String, Object> args = new HashMap();
 
                         args.put("id", id);
@@ -211,36 +214,35 @@ public class MethodCallHandlerImpl implements MethodChannel.MethodCallHandler {
                     public void onDownloadCompleted(String path) {
                         super.onDownloadCompleted(path);
 
-                        final String onDownloadCompleted = helper.call.argument("ondownloadcompleted");
-
-                        if (TextUtils.isEmpty(onDownloadCompleted)) return;
                         Map<String, Object> args = new HashMap();
 
                         args.put("id", id);
                         args.put("path", path);
                         args.put("key", helper.call.argument("key"));
                         channel.invokeMethod("onDownloadCompleted", args);
+
+                        removeTask(id);
                     }
 
                     @Override
                     public void onDownloadError(String errorMessage) {
                         super.onDownloadError(errorMessage);
-                        final String onDownloadError = helper.call.argument("ondownloaderror");
 
-                        if (TextUtils.isEmpty(onDownloadError)) return;
                         Map<String, Object> args = new HashMap();
 
                         args.put("id", id);
                         args.put("error", errorMessage);
                         args.put("key", helper.call.argument("key"));
                         channel.invokeMethod("onDownloadError", args);
+
+                        removeTask(id);
                     }
                 })
                 .build()
                 .startDownloading(context);
     }
 
-    public void removeTask(final long id) {
+    private void removeTask(final long id) {
         tasks.remove(id);
     }