+++ android/src/main/java/com/odehbros/flutter_file_downloader/core/DownloadTask.java
@@ -14,6 +14,7 @@ import android.util.Log;
 import java.util.Map;
 
 import io.flutter.util.PathUtils;
+import com.odehbros.flutter_file_downloader.StoreHelper;
 
 public class DownloadTask {
     final Activity activity;
@@ -22,8 +23,9 @@ public class DownloadTask {
     final Map<String, String> requestHeaders;
     final String notifications;
     private boolean isDownloading = false;
+    final StoreHelper helper;
 
-    public DownloadTask(Activity activity, String url, String name, String notifications, String downloadDestination, DownloadCallbacks callbacks, Map<String, String> requestHeaders) {
+    public DownloadTask(Activity activity, String url, String name, String notifications, String downloadDestination, DownloadCallbacks callbacks, Map<String, String> requestHeaders, StoreHelper helper) {
         this.activity = activity;
         this.url = url;
         this.name = name;
@@ -31,16 +33,30 @@ public class DownloadTask {
         this.downloadDestination = downloadDestination;
         this.callbacks = callbacks;
         this.requestHeaders = requestHeaders;
+        this.helper = helper;
     }
 
     public void startDownloading(final Context context) {
         isDownloading = true;
         final DownloadManager.Request request = new DownloadManager.Request(Uri.parse(url));
 
+        final String downloadFileName = getDownloadedFileName();
+        final String fixedDownloadedFileName = fixDownloadFileName(downloadFileName);
+
+        if (TextUtils.isEmpty(fixedDownloadedFileName)) {
+            final String message = "Invalid file name "+downloadFileName+" -> "+fixedDownloadedFileName+" try changing the download file name";
+            new Handler(Looper.getMainLooper()).post(() -> {
+                callbacks.onDownloadError(message);
+            });
+            helper.result.error("Download file error", message + "", null);
+            return;
+        }
+
+
         if ("appFiles".equals(downloadDestination)) {
-            request.setDestinationInExternalFilesDir(activity, PathUtils.getFilesDir(activity), getDownloadedFileName());
+            request.setDestinationInExternalFilesDir(activity, PathUtils.getFilesDir(activity), fixedDownloadedFileName);
         } else if ("publicDownloads".equals(downloadDestination)) {
-            request.setDestinationInExternalPublicDir(Environment.DIRECTORY_DOWNLOADS, getDownloadedFileName());
+            request.setDestinationInExternalPublicDir(Environment.DIRECTORY_DOWNLOADS, fixedDownloadedFileName);
         }
         if ("disabled".equals(notifications)) {
             request.setNotificationVisibility(DownloadManager.Request.VISIBILITY_HIDDEN);
@@ -50,7 +66,6 @@ public class DownloadTask {
             //DO NOTHING (DEFAULT CASE)
         }
         for (final Map.Entry<String, String> entry : requestHeaders.entrySet()) {
-            System.out.println("HEADER: " + entry.getKey() +":"+ entry.getValue());
             request.addRequestHeader(entry.getKey(), entry.getValue());
         }
         final DownloadManager manager = (DownloadManager) activity.getSystemService(activity.DOWNLOAD_SERVICE);
@@ -61,17 +76,20 @@ public class DownloadTask {
                 trackDownload(manager, downloadedID);
             }
         } catch (Exception e) {
-            Handler uiThreadHandler = new Handler(Looper.getMainLooper());
-            if (e instanceof SecurityException) {
+            String message;
+            if (e.getMessage().startsWith("Unsupported path") || e.getMessage().startsWith("java.io.IOException: Invalid file path")) {
+                message = "Invalid file name "+getDownloadedFileName()+" try changing the download file name";
+            } else if (e instanceof SecurityException) {
                 Log.e("MISSING PERMISSION", "If you want to download a file without notifications, you must provide the permission\n<uses-permission android:name=\"android.permission.DOWNLOAD_WITHOUT_NOTIFICATION\" />");
-                uiThreadHandler.post(() -> {
-                    callbacks.onDownloadError("Missing permission, see the log for more info");
-                });
+                message = "Missing permission, see the log for more info";
             } else {
-                uiThreadHandler.post(() -> {
-                    callbacks.onDownloadError(e.getMessage());
-                });
+                message = e.getMessage();
             }
+
+            new Handler(Looper.getMainLooper()).post(() -> {
+                callbacks.onDownloadError(message);
+            });
+            helper.result.error("Download file error", message + "", null);
         }
     }
 
@@ -141,14 +159,18 @@ public class DownloadTask {
     }
 
     private String getExtension() {
-        String toCheck = name;
-        if (!TextUtils.isEmpty(toCheck)) {
-            if (!toCheck.contains(".")) {
-                final String[] segments = url.split("/");
-                toCheck = segments[segments.length - 1];
-            }
-        } else
-            toCheck = getName();
+//        String toCheck = name;
+//        if (!TextUtils.isEmpty(toCheck)) {
+//            if (!toCheck.contains(".")) {
+//                final String[] segments = url.split("/");
+//                toCheck = segments[segments.length - 1];
+//            }
+//        } else {
+//            toCheck = getName();
+//        }
+
+        final String[] segments = url.split("/");
+        String toCheck = segments[segments.length - 1];
 
         return toCheck.substring(toCheck.lastIndexOf("."));
     }
@@ -159,7 +181,33 @@ public class DownloadTask {
         final String extension = getExtension();
         if (name.contains(".")) {
             name = name.substring(0, name.lastIndexOf("."));
+//            name = fixDownloadFileName(name);
         }
         return String.format("%s.%s", name, extension.replace(".", ""));
+    }
+
+    private String fixDownloadFileName(final String name) {
+        String extension = "";
+        if (name.contains(getExtension())) {
+            extension = name.substring(name.lastIndexOf("."));
+        }
+//        Log.i("FIXING FNAME", name + " -> "+ extension + " | " + getExtension());
+        return name
+                .replace(extension, "")
+                .replace("#", "")
+                .replace("%", "")
+                .replace("*", "")
+                .replace(".", "")
+                .replace("\\", "")
+                .replace("|", "")
+                .replace("\"", "")
+                .replace(":", "")
+                .replace(":", "")
+                .replace("/", "")
+                .replace("<", "")
+                .replace(">", "")
+                .replace("?", "")
+                .replace("&", "")
+                + getExtension();
     }
 }